requirements.txt
mariadb
peewee
requests
pydigitemp